+++ app/src/main/java/org/epstudios/epmobile/DiagnosticScore.kt
@@ -55,21 +55,6 @@ abstract class DiagnosticScore : EpActivity(), View.OnClickListener {
     protected abstract fun init()
 
     protected abstract fun displayResult(message: String?, title: String?)
-//    protected open fun displayResult(message: String?, title: String?) {
-//        // put message in class field so inner class can use
-//        val dialog = AlertDialog.Builder(this).create()
-//        dialog.setMessage(message)
-//        dialog.setButton(
-//            DialogInterface.BUTTON_POSITIVE,
-//            getString(R.string.reset_label),
-//            DialogInterface.OnClickListener { dialog12: DialogInterface?, which: Int -> clearEntries() })
-//        dialog.setButton(
-//            DialogInterface.BUTTON_NEUTRAL,
-//            getString(R.string.dont_reset_label),
-//            DialogInterface.OnClickListener { dialog1: DialogInterface?, which: Int -> })
-//        dialog.setTitle(title)
-//        dialog.show()
-//    }
 
     protected abstract fun clearEntries()
 }